fixup! first pass at integrating chain-validation
