+++ chain/validation/lotus_test.go
@@ -11,29 +11,10 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/filecoin-project/chain-validation/pkg/state"
-	"github.com/filecoin-project/chain-validation/pkg/suites"
 
 	"github.com/filecoin-project/lotus/chain/actors"
 	"github.com/filecoin-project/lotus/chain/address"
 )
-
-func TestStorageMinerValidation(t *testing.T) {
-	factory := NewFactories()
-	suites.CreateStorageMinerAndUpdatePeerIDTest(t, factory)
-
-}
-
-func TestValueTransfer(t *testing.T) {
-	factory := NewFactories()
-	suites.AccountValueTransferSuccess(t, factory, 126)
-	suites.AccountValueTransferZeroFunds(t, factory, 112)
-	suites.AccountValueTransferOverBalanceNonZero(t, factory, 0)
-	suites.AccountValueTransferOverBalanceZero(t, factory, 0)
-	suites.AccountValueTransferToSelf(t, factory, 0)
-	suites.AccountValueTransferFromKnownToUnknownAccount(t, factory, 0)
-	suites.AccountValueTransferFromUnknownToKnownAccount(t, factory, 0)
-	suites.AccountValueTransferFromUnknownToUnknownAccount(t, factory, 0)
-}
 
 //
 // some helpers tests to verify chain-validation can encode parameters to the CBOR lotus uses
